refactor(gui): extract Logs and Maintenance behavior into mixins
Move two cohesive blocks off WinpodxWindow into private mixin
modules so main_window.py can focus on overall window orchestration:

- LogsMixin (src/winpodx/gui/_main_window_logs.py)
  Logs-tab behavior: _log_append, _run_log_cmd, _on_follow_pod_log,
  _on_tail_app_log, _on_follow_app_log, _drain_tail, _on_stop_tail,
  _on_cmd_enter, and the _ALLOWED_COMMANDS allowlist.

- MaintenanceMixin (src/winpodx/gui/_main_window_maintenance.py)
  Maintenance-tab button handlers: _on_cleanup, _refresh_update_status,
  _on_enable_updates, _on_disable_updates, _on_timesync, _on_suspend,
  _on_resume, _on_debloat, _on_apply_fixes, _on_open_desktop.

Mixins are intentionally not standalone — they document but don't
enforce the host-class attribute contract (log_output, info_label,
signals, etc.). WinpodxWindow now reads
`class WinpodxWindow(LogsMixin, MaintenanceMixin, QMainWindow)`.

No behavior change; all call sites unaffected (the methods are still
attributes of WinpodxWindow via MRO). Tests: 1096 passed.

main_window.py: 2545 -> 2255 lines (-290) across both slices.

Follow-up to the previous slice (static widget helpers). Subsequent
slices will pull out the _build_* UI builders (the biggest remaining
chunk).
