+++ src/winpodx/gui/_main_window_maintenance.py
@@ -1,0 +1,230 @@
+"""Maintenance-tab mixin for ``WinpodxWindow``.
+
+Holds the slot handlers driven by Maintenance-page buttons: lock-file
+cleanup, Windows Update enable/disable, time sync, suspend / resume,
+debloat, Windows-side runtime fixes, and "open Windows desktop". Pulled
+out of ``main_window.py`` to keep that file focused on overall window
+orchestration.
+
+Host-class contract (only listed for readers; not enforced):
+    info_label: QLabel              — the small status text below buttons.
+    _update_status_label: QLabel    — Windows-Update status line.
+    _btn_enable_updates: QPushButton
+    _btn_disable_updates: QPushButton
+    app_launched: Signal(str)
+    app_launch_failed: Signal(str)
+    pod_status_updated: Signal(str, str)
+    _refresh_pod_status() -> None   — defined on the host class.
+"""
+
+from __future__ import annotations
+
+import threading
+
+from PySide6.QtWidgets import QMessageBox
+
+from winpodx.core.config import Config
+from winpodx.utils.paths import bundle_dir
+
+
+class MaintenanceMixin:
+    """Maintenance-tab behavior. Mix into ``WinpodxWindow``."""
+
+    def _on_cleanup(self) -> None:
+        from winpodx.core.daemon import cleanup_lock_files
+
+        removed = cleanup_lock_files()
+        msg = f"Removed {len(removed)} lock files" if removed else "No lock files found"
+        self.info_label.setText(msg)
+
+    def _refresh_update_status(self) -> None:
+        def _do() -> None:
+            from winpodx.core.updates import get_update_status
+
+            cfg = Config.load()
+            status = get_update_status(cfg)
+            if status == "enabled":
+                self._update_status_label.setText("Windows Update is enabled")
+                self._btn_enable_updates.setEnabled(False)
+                self._btn_disable_updates.setEnabled(True)
+            elif status == "disabled":
+                self._update_status_label.setText("Windows Update is disabled")
+                self._btn_enable_updates.setEnabled(True)
+                self._btn_disable_updates.setEnabled(False)
+            else:
+                self._update_status_label.setText("Status unknown (container not running?)")
+                self._btn_enable_updates.setEnabled(True)
+                self._btn_disable_updates.setEnabled(True)
+
+        threading.Thread(target=_do, daemon=True).start()
+
+    def _on_enable_updates(self) -> None:
+        self._update_status_label.setText("Enabling Windows Update...")
+        self._btn_enable_updates.setEnabled(False)
+        self._btn_disable_updates.setEnabled(False)
+
+        def _do() -> None:
+            from winpodx.core.updates import enable_updates
+
+            cfg = Config.load()
+            ok = enable_updates(cfg)
+            if ok:
+                self.app_launched.emit("Windows Update enabled")
+            else:
+                self.app_launch_failed.emit("Failed to enable Windows Update")
+            self._refresh_update_status()
+
+        threading.Thread(target=_do, daemon=True).start()
+
+    def _on_disable_updates(self) -> None:
+        reply = QMessageBox.question(
+            self,
+            "Disable Windows Update",
+            "This will stop Windows Update services and block update domains.\n"
+            "Security updates will NOT be installed while disabled.\n\nProceed?",
+            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
+            QMessageBox.StandardButton.No,
+        )
+        if reply != QMessageBox.StandardButton.Yes:
+            return
+
+        self._update_status_label.setText("Disabling Windows Update...")
+        self._btn_enable_updates.setEnabled(False)
+        self._btn_disable_updates.setEnabled(False)
+
+        def _do() -> None:
+            from winpodx.core.updates import disable_updates
+
+            cfg = Config.load()
+            ok = disable_updates(cfg)
+            if ok:
+                self.app_launched.emit("Windows Update disabled")
+            else:
+                self.app_launch_failed.emit("Failed to disable Windows Update")
+            self._refresh_update_status()
+
+        threading.Thread(target=_do, daemon=True).start()
+
+    def _on_timesync(self) -> None:
+        from winpodx.core.daemon import sync_windows_time
+
+        ok = sync_windows_time(Config.load())
+        self.info_label.setText("Time synced" if ok else "Time sync failed")
+
+    def _on_suspend(self) -> None:
+        from winpodx.core.daemon import suspend_pod
+
+        ok = suspend_pod(Config.load())
+        self.info_label.setText("Pod suspended" if ok else "Suspend failed")
+        self._refresh_pod_status()
+
+    def _on_resume(self) -> None:
+        from winpodx.core.daemon import resume_pod
+
+        ok = resume_pod(Config.load())
+        self.info_label.setText("Pod resumed" if ok else "Resume failed")
+        self._refresh_pod_status()
+
+    def _on_debloat(self) -> None:
+        reply = QMessageBox.question(
+            self,
+            "Debloat",
+            "This will disable telemetry, ads, and bloat in Windows.\nProceed?",
+        )
+        if reply != QMessageBox.StandardButton.Yes:
+            return
+
+        self.info_label.setText("Running debloat...")
+
+        def _do() -> None:
+            from winpodx.core.windows_exec import WindowsExecError, run_via_transport
+
+            cfg = Config.load()
+            script = bundle_dir() / "scripts" / "windows" / "debloat.ps1"
+            if not script.exists():
+                self.app_launch_failed.emit("Debloat script not found")
+                return
+
+            try:
+                payload = script.read_text(encoding="utf-8")
+            except OSError as e:
+                self.app_launch_failed.emit(f"Cannot read debloat script: {e}")
+                return
+
+            try:
+                result = run_via_transport(cfg, payload, description="debloat", timeout=180)
+            except WindowsExecError as e:
+                self.app_launch_failed.emit(f"Debloat channel failure: {e}")
+                return
+
+            if result.rc == 0:
+                self.app_launched.emit("Debloat complete")
+            else:
+                self.app_launch_failed.emit(
+                    f"Debloat failed (rc={result.rc}): "
+                    f"{result.stderr.strip() or result.stdout.strip()[:200]}"
+                )
+            self.pod_status_updated.emit("running", cfg.rdp.ip)
+
+        threading.Thread(target=_do, daemon=True).start()
+
+    def _on_apply_fixes(self) -> None:
+        """v0.1.9.3: Apply Windows-side runtime fixes to the existing pod.
+
+        Same idempotent helpers fired by `winpodx pod apply-fixes` and by
+        `provisioner.ensure_ready` — but on demand from the GUI for users
+        whose migrate short-circuited with "already current" so the
+        Windows VM never received the OEM v7+v8 fixes.
+        """
+        self.info_label.setText("Applying Windows-side fixes...")
+
+        def _do() -> None:
+            from winpodx.core.pod import PodState, pod_status
+            from winpodx.core.provisioner import apply_windows_runtime_fixes
+
+            cfg = Config.load()
+            try:
+                state = pod_status(cfg).state
+            except Exception as e:  # noqa: BLE001
+                self.app_launch_failed.emit(f"Apply fixes failed (pod probe): {e}")
+                return
+
+            if state != PodState.RUNNING:
+                self.app_launch_failed.emit(
+                    "Pod is not running — start it first via the Apps page or "
+                    "`winpodx pod start --wait`."
+                )
+                return
+
+            try:
+                results = apply_windows_runtime_fixes(cfg)
+            except Exception as e:  # noqa: BLE001
+                self.app_launch_failed.emit(f"Apply fixes raised: {e}")
+                return
+
+            ok_count = sum(1 for v in results.values() if v == "ok")
+            total = len(results)
+            failed = [k for k, v in results.items() if v != "ok"]
+            if failed:
+                detail = ", ".join(failed)
+                self.app_launch_failed.emit(f"Apply fixes: {ok_count}/{total} OK; failed: {detail}")
+            else:
+                self.app_launched.emit(f"Windows-side fixes applied ({ok_count}/{total} OK)")
+
+        threading.Thread(target=_do, daemon=True).start()
+
+    def _on_open_desktop(self) -> None:
+        self.info_label.setText("Opening Windows desktop...")
+
+        def _do() -> None:
+            try:
+                from winpodx.core.provisioner import ensure_ready
+                from winpodx.core.rdp import launch_desktop
+
+                cfg = ensure_ready()
+                launch_desktop(cfg)
+                self.app_launched.emit("Windows Desktop")
+            except Exception as e:  # noqa: BLE001
+                self.app_launch_failed.emit(str(e))
+
+        threading.Thread(target=_do, daemon=True).start()
