+++ src/winpodx/gui/main_window.py
@@ -32,6 +32,8 @@ from PySide6.QtWidgets import (
 from winpodx.core.app import AppInfo, list_available_apps
 from winpodx.core.config import Config
 from winpodx.core.pod import pod_status
+from winpodx.gui._main_window_logs import LogsMixin
+from winpodx.gui._main_window_maintenance import MaintenanceMixin
 from winpodx.gui._widget_helpers import add_shadow, make_app_avatar, make_source_badge
 from winpodx.gui.theme import (
     ACTION_ROW,
@@ -61,12 +63,11 @@ from winpodx.gui.theme import (
     avatar_color,
 )
 from winpodx.gui.workers import DiscoveryWorker, InfoWorker
-from winpodx.utils.paths import bundle_dir
 
 log = logging.getLogger(__name__)
 
 
-class WinpodxWindow(QMainWindow):
+class WinpodxWindow(LogsMixin, MaintenanceMixin, QMainWindow):
     """Main window with horizontal top navigation bar."""
 
     # Thread-safe signals
@@ -1566,194 +1567,6 @@ class WinpodxWindow(QMainWindow):
             cfg_rows.append(("WARNING", warning))
         self._set_info_card_rows("config", cfg_rows)
 
-    def _log_append(self, text: str, color: str = C.SUBTEXT1) -> None:
-        """Append colored text to the log output."""
-        import html
-
-        safe = html.escape(text)
-        self.log_output.append(f'<span style="color:{color}">{safe}</span>')
-        sb = self.log_output.verticalScrollBar()
-        sb.setValue(sb.maximum())
-
-    def _run_log_cmd(self, cmd: list[str]) -> None:
-        """Run command and show output in terminal."""
-        import subprocess
-
-        self._log_append(f"$ {' '.join(cmd)}", C.BLUE)
-
-        def _do() -> None:
-            try:
-                result = subprocess.run(cmd, capture_output=True, text=True, timeout=30)
-                if result.stdout.strip():
-                    self.log_signal.emit(result.stdout.strip(), C.SUBTEXT1)
-                if result.stderr.strip():
-                    self.log_signal.emit(result.stderr.strip(), C.YELLOW)
-                if result.returncode != 0:
-                    self.log_signal.emit(f"Exit code: {result.returncode}", C.RED)
-            except subprocess.TimeoutExpired:
-                self.log_signal.emit("Command timed out (30s)", C.RED)
-            except FileNotFoundError:
-                self.log_signal.emit(f"Command not found: {cmd[0]}", C.RED)
-
-        threading.Thread(target=_do, daemon=True).start()
-
-    # v0.2.0.10: live log streaming. The Pod logs button shows the last
-    # 100 lines, but for first-install / debug the user wants to watch
-    # the container output as Windows downloads / Sysprep / boots, and
-    # also see winpodx's own application log (under XDG state) so they
-    # can correlate guest events with host actions.
-    def _on_follow_pod_log(self) -> None:
-        import subprocess
-
-        self._on_stop_tail()
-        self._log_append(
-            f"$ podman logs -f --tail 50 {self.cfg.pod.container_name} (Stop tail to end)",
-            C.BLUE,
-        )
-        try:
-            self._tail_proc = subprocess.Popen(
-                [
-                    self.cfg.pod.backend,
-                    "logs",
-                    "-f",
-                    "--tail",
-                    "50",
-                    self.cfg.pod.container_name,
-                ],
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-                bufsize=1,
-            )
-        except (FileNotFoundError, OSError) as e:
-            self._log_append(f"Could not start tail: {e}", C.RED)
-            return
-        self._tail_stop = threading.Event()
-        threading.Thread(target=self._drain_tail, args=(self._tail_proc,), daemon=True).start()
-
-    def _on_tail_app_log(self) -> None:
-        from winpodx.utils.paths import config_dir
-
-        log_path = config_dir() / "winpodx.log"
-        self._log_append(f"$ tail {log_path}", C.BLUE)
-        if not log_path.exists():
-            self._log_append(
-                "(no app log file yet — winpodx writes to it after the next CLI / GUI action)",
-                C.OVERLAY0,
-            )
-            return
-        try:
-            content = log_path.read_text(encoding="utf-8", errors="replace")
-        except OSError as e:
-            self._log_append(f"Could not read app log: {e}", C.RED)
-            return
-        # Show only the last ~200 lines so we don't drown the view.
-        lines = content.splitlines()[-200:]
-        for line in lines:
-            self._log_append(line, C.SUBTEXT1)
-
-    def _on_follow_app_log(self) -> None:
-        import subprocess
-
-        from winpodx.utils.paths import config_dir
-
-        log_path = config_dir() / "winpodx.log"
-        self._on_stop_tail()
-        self._log_append(f"$ tail -F {log_path} (Stop tail to end)", C.BLUE)
-        # Pre-create the file so `tail -F` doesn't loop on FileNotFoundError.
-        log_path.parent.mkdir(parents=True, exist_ok=True)
-        log_path.touch(exist_ok=True)
-        try:
-            self._tail_proc = subprocess.Popen(
-                ["tail", "-F", "-n", "50", str(log_path)],
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-                bufsize=1,
-            )
-        except (FileNotFoundError, OSError) as e:
-            self._log_append(f"Could not start tail: {e}", C.RED)
-            return
-        self._tail_stop = threading.Event()
-        threading.Thread(target=self._drain_tail, args=(self._tail_proc,), daemon=True).start()
-
-    def _drain_tail(self, proc) -> None:  # type: ignore[no-untyped-def]
-        try:
-            for line in iter(proc.stdout.readline, ""):
-                if self._tail_stop.is_set():
-                    break
-                line = line.rstrip()
-                if line:
-                    self.log_signal.emit(line, C.SUBTEXT1)
-        except Exception:  # noqa: BLE001
-            log.debug("tail drain crashed", exc_info=True)
-        try:
-            proc.terminate()
-            proc.wait(timeout=2)
-        except Exception:  # noqa: BLE001
-            try:
-                proc.kill()
-            except Exception:  # noqa: BLE001
-                pass
-
-    def _on_stop_tail(self) -> None:
-        proc = getattr(self, "_tail_proc", None)
-        stop = getattr(self, "_tail_stop", None)
-        if proc is None:
-            return
-        if stop is not None:
-            stop.set()
-        try:
-            proc.terminate()
-            proc.wait(timeout=2)
-        except Exception:  # noqa: BLE001
-            try:
-                proc.kill()
-            except Exception:  # noqa: BLE001
-                pass
-        self._log_append("(tail stopped)", C.OVERLAY0)
-        self._tail_proc = None
-
-    _ALLOWED_COMMANDS = {
-        "podman",
-        "docker",
-        "virsh",
-        "winpodx",
-        "podman-compose",
-        "docker-compose",
-        "xfreerdp",
-        "xfreerdp3",
-        "wlfreerdp",
-        "wlfreerdp3",
-        "systemctl",
-        "journalctl",
-        "ss",
-        "ip",
-        "ping",
-    }
-
-    def _on_cmd_enter(self) -> None:
-        """Handle command input (allowlist-based)."""
-        import shlex
-
-        text = self.cmd_input.text().strip()
-        if not text:
-            return
-        self.cmd_input.clear()
-
-        try:
-            cmd = shlex.split(text)
-        except ValueError as e:
-            self._log_append(f"Parse error: {e}", C.RED)
-            return
-
-        if not cmd or cmd[0] not in self._ALLOWED_COMMANDS:
-            allowed = ", ".join(sorted(self._ALLOWED_COMMANDS))
-            self._log_append(f"Blocked: allowed commands: {allowed}", C.RED)
-            return
-
-        self._run_log_cmd(cmd)
-
     def _on_rdp_test(self) -> None:
         self._log_append("$ Testing RDP connection...", C.BLUE)
 
@@ -2191,205 +2004,6 @@ class WinpodxWindow(QMainWindow):
                 return
 
         self.info_label.setText("Settings saved")
-
-    def _on_cleanup(self) -> None:
-        from winpodx.core.daemon import cleanup_lock_files
-
-        removed = cleanup_lock_files()
-        msg = f"Removed {len(removed)} lock files" if removed else "No lock files found"
-        self.info_label.setText(msg)
-
-    def _refresh_update_status(self) -> None:
-        def _do() -> None:
-            from winpodx.core.updates import get_update_status
-
-            cfg = Config.load()
-            status = get_update_status(cfg)
-            if status == "enabled":
-                self._update_status_label.setText("Windows Update is enabled")
-                self._btn_enable_updates.setEnabled(False)
-                self._btn_disable_updates.setEnabled(True)
-            elif status == "disabled":
-                self._update_status_label.setText("Windows Update is disabled")
-                self._btn_enable_updates.setEnabled(True)
-                self._btn_disable_updates.setEnabled(False)
-            else:
-                self._update_status_label.setText("Status unknown (container not running?)")
-                self._btn_enable_updates.setEnabled(True)
-                self._btn_disable_updates.setEnabled(True)
-
-        threading.Thread(target=_do, daemon=True).start()
-
-    def _on_enable_updates(self) -> None:
-        self._update_status_label.setText("Enabling Windows Update...")
-        self._btn_enable_updates.setEnabled(False)
-        self._btn_disable_updates.setEnabled(False)
-
-        def _do() -> None:
-            from winpodx.core.updates import enable_updates
-
-            cfg = Config.load()
-            ok = enable_updates(cfg)
-            if ok:
-                self.app_launched.emit("Windows Update enabled")
-            else:
-                self.app_launch_failed.emit("Failed to enable Windows Update")
-            self._refresh_update_status()
-
-        threading.Thread(target=_do, daemon=True).start()
-
-    def _on_disable_updates(self) -> None:
-        reply = QMessageBox.question(
-            self,
-            "Disable Windows Update",
-            "This will stop Windows Update services and block update domains.\n"
-            "Security updates will NOT be installed while disabled.\n\nProceed?",
-            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
-            QMessageBox.StandardButton.No,
-        )
-        if reply != QMessageBox.StandardButton.Yes:
-            return
-
-        self._update_status_label.setText("Disabling Windows Update...")
-        self._btn_enable_updates.setEnabled(False)
-        self._btn_disable_updates.setEnabled(False)
-
-        def _do() -> None:
-            from winpodx.core.updates import disable_updates
-
-            cfg = Config.load()
-            ok = disable_updates(cfg)
-            if ok:
-                self.app_launched.emit("Windows Update disabled")
-            else:
-                self.app_launch_failed.emit("Failed to disable Windows Update")
-            self._refresh_update_status()
-
-        threading.Thread(target=_do, daemon=True).start()
-
-    def _on_timesync(self) -> None:
-        from winpodx.core.daemon import sync_windows_time
-
-        ok = sync_windows_time(Config.load())
-        self.info_label.setText("Time synced" if ok else "Time sync failed")
-
-    def _on_suspend(self) -> None:
-        from winpodx.core.daemon import suspend_pod
-
-        ok = suspend_pod(Config.load())
-        self.info_label.setText("Pod suspended" if ok else "Suspend failed")
-        self._refresh_pod_status()
-
-    def _on_resume(self) -> None:
-        from winpodx.core.daemon import resume_pod
-
-        ok = resume_pod(Config.load())
-        self.info_label.setText("Pod resumed" if ok else "Resume failed")
-        self._refresh_pod_status()
-
-    def _on_debloat(self) -> None:
-        reply = QMessageBox.question(
-            self,
-            "Debloat",
-            "This will disable telemetry, ads, and bloat in Windows.\nProceed?",
-        )
-        if reply != QMessageBox.StandardButton.Yes:
-            return
-
-        self.info_label.setText("Running debloat...")
-
-        def _do() -> None:
-            from winpodx.core.windows_exec import WindowsExecError, run_via_transport
-
-            cfg = Config.load()
-            script = bundle_dir() / "scripts" / "windows" / "debloat.ps1"
-            if not script.exists():
-                self.app_launch_failed.emit("Debloat script not found")
-                return
-
-            try:
-                payload = script.read_text(encoding="utf-8")
-            except OSError as e:
-                self.app_launch_failed.emit(f"Cannot read debloat script: {e}")
-                return
-
-            try:
-                result = run_via_transport(cfg, payload, description="debloat", timeout=180)
-            except WindowsExecError as e:
-                self.app_launch_failed.emit(f"Debloat channel failure: {e}")
-                return
-
-            if result.rc == 0:
-                self.app_launched.emit("Debloat complete")
-            else:
-                self.app_launch_failed.emit(
-                    f"Debloat failed (rc={result.rc}): "
-                    f"{result.stderr.strip() or result.stdout.strip()[:200]}"
-                )
-            self.pod_status_updated.emit("running", cfg.rdp.ip)
-
-        threading.Thread(target=_do, daemon=True).start()
-
-    def _on_apply_fixes(self) -> None:
-        """v0.1.9.3: Apply Windows-side runtime fixes to the existing pod.
-
-        Same idempotent helpers fired by `winpodx pod apply-fixes` and by
-        `provisioner.ensure_ready` — but on demand from the GUI for users
-        whose migrate short-circuited with "already current" so the
-        Windows VM never received the OEM v7+v8 fixes.
-        """
-        self.info_label.setText("Applying Windows-side fixes...")
-
-        def _do() -> None:
-            from winpodx.core.pod import PodState, pod_status
-            from winpodx.core.provisioner import apply_windows_runtime_fixes
-
-            cfg = Config.load()
-            try:
-                state = pod_status(cfg).state
-            except Exception as e:  # noqa: BLE001
-                self.app_launch_failed.emit(f"Apply fixes failed (pod probe): {e}")
-                return
-
-            if state != PodState.RUNNING:
-                self.app_launch_failed.emit(
-                    "Pod is not running — start it first via the Apps page or "
-                    "`winpodx pod start --wait`."
-                )
-                return
-
-            try:
-                results = apply_windows_runtime_fixes(cfg)
-            except Exception as e:  # noqa: BLE001
-                self.app_launch_failed.emit(f"Apply fixes raised: {e}")
-                return
-
-            ok_count = sum(1 for v in results.values() if v == "ok")
-            total = len(results)
-            failed = [k for k, v in results.items() if v != "ok"]
-            if failed:
-                detail = ", ".join(failed)
-                self.app_launch_failed.emit(f"Apply fixes: {ok_count}/{total} OK; failed: {detail}")
-            else:
-                self.app_launched.emit(f"Windows-side fixes applied ({ok_count}/{total} OK)")
-
-        threading.Thread(target=_do, daemon=True).start()
-
-    def _on_open_desktop(self) -> None:
-        self.info_label.setText("Opening Windows desktop...")
-
-        def _do() -> None:
-            try:
-                from winpodx.core.provisioner import ensure_ready
-                from winpodx.core.rdp import launch_desktop
-
-                cfg = ensure_ready()
-                launch_desktop(cfg)
-                self.app_launched.emit("Windows Desktop")
-            except Exception as e:
-                self.app_launch_failed.emit(str(e))
-
-        threading.Thread(target=_do, daemon=True).start()
 
     def _maybe_run_first_launch_checks(self) -> None:
         """v0.2.1: on GUI startup, resume any pending install steps and —
